Crate thread_id

Thread-ID: get a unique ID for the current thread.

For diagnostics and debugging it can often be useful to get an ID that is different for every thread. This crate provides that functionality.

Example

use std::thread;
use thread_id;

thread::spawn(move || {
    println!("spawned thread has id {}", thread_id::get());
});

println!("main thread has id {}", thread_id::get());

Functions

get

Returns a number that is unique to the calling thread.